Bug fixes, added clean rooms job, added delete current account, minor chages., docker compose updated

+++ src/backend/StreamWatch.Core/Options/S3StorageOptions.cs
@@ -7,5 +7,5 @@ public class S3StorageOptions
     public string ServiceURL { get; set; } = "";
     public string PublicBucket { get; set; } = "";
     public string PublicBaseUrl { get; set; } = "";
-    public int LimitUsagePerUser {get; set;} = 0; //
+    public long LimitUsagePerUser {get; set;} = 0; //
 }
